Add 'Percentage' example to documentation
A new example named 'Percentage' has been added to the Graphane documentation. The related markdown file has been created and the link to this new example has been included in the configuration file for VitePress. This new example will help users understand the usage of the 'Percentage' component in the Graphane library.