Update Graphane examples with inline editor links
The changes update several Graphane example pages to include inline editor links. Majority of the updates involved the removal of the style="width:100%" attribute from g-editor elements and marking the start of examples with an enclosing <div id="example"> to facilitate referencing in the g-editor href field. Also, redundant code snippets displayed beneath the live examples were removed in some of the documents. For the Du Bois Tribute example, a large part of the SVG code was modified to condense the presentation.